third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/open-url-encoding-expected.txt

This is a testharness.js-based test.
FAIL percent encode characters assert_equals: expected "%C3%9F" but got "%DF"
FAIL lone surrogate should return U+FFFD assert_equals: expected "%EF%BF%BD" but got "%26%2355357%3B"
Harness: the test ran to completion.
